Added where user can go
Added where user can go when he took piece and some refactoring
