+++ Assets/Scripts/ChessFigures/Figure.cs
@@ -18,7 +18,7 @@ public abstract class Figure
     {
         List<Point> result = new();
 
-        for (int x = x0 + sx, y = y0 + sy; GetMap(x, y, map) == 0; x += sx, y += sy)
+        for (int x = x0 + sx, y = y0 + sy; Map.GetMap(x, y, map) == 0; x += sx, y += sy)
         {
             result.Add(new Point(x, y));
         }
@@ -29,11 +29,11 @@ public abstract class Figure
     public List<Point> FindConnectedPiecesForLongRangeFigures(int x0, int y0, int sx, int sy, int[,] map)
     {
         List<Point> result = new();
-        int piece = GetMap(x0, y0, map);
+        int piece = Map.GetMap(x0, y0, map);
 
-        for (int x = x0 + sx, y = y0 + sy; (GetMap(x, y, map) == 0) || (GetMap(x, y, map) == piece); x += sx, y += sy)
+        for (int x = x0 + sx, y = y0 + sy; (Map.GetMap(x, y, map) == 0) || (Map.GetMap(x, y, map) == piece); x += sx, y += sy)
         {
-            if (GetMap(x, y, map) == piece)
+            if (Map.GetMap(x, y, map) == piece)
             {
                 result.Add(new Point(x, y));
                 break;
@@ -41,17 +41,6 @@ public abstract class Figure
         }
 
         return result;
-    }
-
-    public bool OnMap(int x, int y)
-    {
-        return x >= 0 && x < Board.SIZE && y >= 0 && y < Board.SIZE;
-    }
-
-    public int GetMap(int x, int y, int[,] map)
-    {
-        if (!OnMap(x, y)) return -1;
-        return map[x, y];
     }
 
     public abstract bool CanMove(Point toLocation, int[,] map);
